Moves pages into modules
