--- features/page_objects/external/returns/nil_return_page.rb
@@ -1,13 +1,0 @@
-require_relative "../../sections/error_summary.rb"
-
-class ExternalReturnsNilReturnPage < SitePrism::Page
-
-  set_url "#{external_application_url}return/confirm{?returnId}"
-  set_url_matcher %r{\/return\/confirm\?returnId=}
-
-  element(:licence_number, ".govuk-caption-l")
-  element(:heading, "h1.govuk-heading-l")
-  element(:submit_button, "button[type=submit]")
-
-  section(:error_summary, ErrorSummarySection, ".govuk-error-summary")
-end
