--- features/page_objects/external_account_settings_page.rb
@@ -1,8 +1,0 @@
-require "uri"
-
-class ExternalAccountSettingsPage < SitePrism::Page
-
-  account_page_url = URI.join(external_application_url, "account/")
-  set_url(account_page_url)
-
-end
