+++ features/support/helpers.rb
@@ -78,9 +78,9 @@ def config_accounts(key)
 end
 
 def external_returns_page_from_question(text)
-  return ExternalReturnsHasWaterBeenAbstractedPage.new if text == "Have you abstracted water"
-  return ExternalReturnsHowAreYouReportingFiguresPage.new if text == "How are you reporting your figures"
-  return ExternalReturnsNilReturnPage.new if text == "Nil return"
+  return Pages::External::Returns::HasWaterBeenAbstracted.new if text == "Have you abstracted water"
+  return Pages::External::Returns::HowAreYouReportingFigures.new if text == "How are you reporting your figures"
+  return Pages::External::Returns::NilReturn.new if text == "Nil return"
 
   raise "Cannot resolve question text to page: #{text}"
 end
